agent_service/requirements.txt
# =============================
# requirements.txt for agent_service
# Location: KIFrag/agent_service/requirements.txt
# Description: Python dependencies for the Agent Service (FastAPI app for agent logic)
# =============================
fastapi
uvicorn
langchain
langgraph
python-dotenv
langchain_community
pinecone
pinecone-client
langchain_pinecone
langchain_huggingface
langchain-together
duckduckgo-search
sentence-transformers